+++ .rtd_pip_reqs.txt
@@ -3,4 +3,3 @@ nbsphinx>=0.6.0
 recommonmark>=0.6.0
 sphinx_rtd_theme>=0.4.3
 sphinx>=3.0.4
-theano>=1.0.4
